feat: send bitcoin back navigation
