+++ app/screens/send-bitcoin-screen/send-bitcoin-details-screen.tsx
@@ -1,0 +1,639 @@
+import useMainQuery from "@app/hooks/use-main-query"
+import React, { useCallback, useEffect, useState } from "react"
+import {
+  ScrollView,
+  StyleSheet,
+  Text,
+  TextInput,
+  TouchableWithoutFeedback,
+  View,
+} from "react-native"
+import { palette } from "@app/theme"
+import { PaymentAmount, WalletCurrency } from "@app/types/amounts"
+import { fetchLnurlInvoice, translateUnknown as translate } from "@galoymoney/client"
+import { Satoshis } from "lnurl-pay/dist/types/types"
+import { StackScreenProps } from "@react-navigation/stack"
+import { RootStackParamList } from "@app/navigation/stack-param-lists"
+import { useMySubscription, useWalletBalance } from "@app/hooks"
+import {
+  paymentAmountToDollarsOrSats,
+  satAmountDisplay,
+  usdAmountDisplay,
+} from "@app/utils/currencyConversion"
+import ReactNativeModal from "react-native-modal"
+import { FakeCurrencyInput } from "react-native-currency-input"
+import SwitchIcon from "@app/assets/icons/switch.svg"
+import Icon from "react-native-vector-icons/Ionicons"
+import NoteIcon from "@app/assets/icons/note.svg"
+import { Button } from "react-native-elements"
+
+const Styles = StyleSheet.create({
+  scrollView: {
+    flexDirection: "column",
+    padding: 20,
+    flex: 6,
+  },
+  contentContainer: {
+    flexGrow: 1,
+  },
+  sendBitcoinAmountContainer: {
+    flex: 1,
+  },
+  fieldBackground: {
+    flexDirection: "row",
+    borderStyle: "solid",
+    overflow: "hidden",
+    backgroundColor: palette.white,
+    paddingHorizontal: 14,
+    marginBottom: 12,
+    borderRadius: 10,
+    alignItems: "center",
+    height: 60,
+  },
+  walletSelectorTypeContainer: {
+    justifyContent: "center",
+    alignItems: "flex-start",
+    width: 50,
+    marginRight: 20,
+  },
+  walletSelectorTypeLabelBitcoin: {
+    height: 30,
+    width: 50,
+    borderRadius: 10,
+    backgroundColor: "rgba(241, 164, 60, 0.5)",
+    justifyContent: "center",
+    alignItems: "center",
+  },
+  walletSelectorTypeLabelUsd: {
+    height: 30,
+    width: 50,
+    backgroundColor: palette.usdSecondary,
+    borderRadius: 10,
+    justifyContent: "center",
+    alignItems: "center",
+  },
+  walletSelectorTypeLabelUsdText: {
+    fontWeight: "bold",
+    color: palette.usdPrimary,
+  },
+  walletSelectorTypeLabelBtcText: {
+    fontWeight: "bold",
+    color: palette.btcPrimary,
+  },
+  walletSelectorInfoContainer: {
+    flex: 1,
+    flexDirection: "column",
+  },
+  walletTitleContainer: {
+    flex: 1,
+  },
+  walletBalanceContainer: {
+    flex: 1,
+  },
+  walletTypeText: {
+    fontWeight: "bold",
+    fontSize: 18,
+    color: palette.lapisLazuli,
+  },
+  walletSelectorTypeTextContainer: {
+    flex: 1,
+    justifyContent: "flex-end",
+  },
+  walletSelectorBalanceContainer: {
+    flex: 1,
+    flexDirection: "row",
+  },
+  walletBalanceText: {
+    color: palette.midGrey,
+  },
+  fieldTitleText: {
+    fontWeight: "bold",
+    color: palette.lapisLazuli,
+    marginBottom: 4,
+  },
+  fieldContainer: {},
+  currencyInputContainer: {
+    flexDirection: "column",
+    flex: 1,
+  },
+  switchCurrencyIconContainer: {
+    width: 50,
+    justifyContent: "center",
+    alignItems: "center",
+  },
+  walletBalanceInput: {
+    color: palette.lapisLazuli,
+    fontSize: 20,
+    fontWeight: "600",
+  },
+  convertedAmountText: {
+    color: palette.coolGrey,
+    fontSize: 12,
+  },
+  errorContainer: {
+    marginVertical: 20,
+    flex: 1,
+  },
+  errorText: {
+    color: palette.red,
+    textAlign: "center",
+  },
+  noteContainer: {
+    flex: 1,
+    flexDirection: "row",
+  },
+  noteIconContainer: {
+    marginRight: 12,
+    justifyContent: "center",
+    alignItems: "flex-start",
+  },
+  noteIcon: {
+    justifyContent: "center",
+    alignItems: "center",
+  },
+  noteInput: {
+    flex: 1,
+  },
+  button: {
+    marginTop: 20,
+    height: 60,
+    borderRadius: 10,
+  },
+  disabledButtonStyle: {
+    backgroundColor: "rgba(83, 111, 242, 0.1)",
+  },
+  disabledButtonTitleStyle: {
+    color: palette.lightBlue,
+    fontWeight: "600",
+  },
+  activeButtonStyle: {
+    backgroundColor: palette.lightBlue,
+  },
+  activeButtonTitleStyle: {
+    color: palette.white,
+    fontWeight: "bold",
+  },
+  buttonContainer: {
+    flex: 1,
+    justifyContent: "flex-end",
+    marginBottom: 50,
+  },
+  modal: {
+    marginBottom: "90%",
+  },
+  pickWalletIcon: {
+    marginRight: 12,
+  },
+})
+
+const SendBitcoinDetailsScreen = ({
+  navigation,
+  route,
+}: StackScreenProps<RootStackParamList, "sendBitcoinDetails">) => {
+  const {
+    fixedAmount,
+    destination,
+    note: initialNote,
+    lnurl: lnurlParams,
+    recipientWalletId,
+    paymentType,
+    sameNode,
+  } = route.params
+
+  const { defaultWallet, usdWalletId } = useMainQuery()
+  const [amountCurrency, setAmountCurrency] = useState<WalletCurrency>(
+    fixedAmount ? WalletCurrency.BTC : WalletCurrency.USD,
+  )
+  const [note, setNote] = useState(initialNote)
+  const [fromWallet, setFromWallet] = useState(defaultWallet)
+
+  const toggleAmountCurrency = useCallback(() => {
+    if (fixedAmount) {
+      return
+    }
+
+    if (amountCurrency === WalletCurrency.USD) {
+      setAmountCurrency(WalletCurrency.BTC)
+    }
+    if (amountCurrency === WalletCurrency.BTC) {
+      setAmountCurrency(WalletCurrency.USD)
+    }
+  }, [amountCurrency, fixedAmount])
+
+  const [isModalVisible, setIsModalVisible] = useState(false)
+  const { wallets } = useMainQuery()
+  const { usdWalletBalance, btcWalletBalance, btcWalletValueInUsd } = useWalletBalance()
+  const [centAmount, setCentAmount] = useState<PaymentAmount<WalletCurrency.USD>>({
+    amount: 0,
+    currency: WalletCurrency.USD,
+  })
+  const [satAmount, setSatAmount] = useState<PaymentAmount<WalletCurrency.BTC>>(
+    fixedAmount || { amount: 0, currency: WalletCurrency.BTC },
+  )
+  const setRawCentAmount = (dollars) => {
+    setCentAmount({ amount: dollars * 100, currency: WalletCurrency.USD })
+  }
+  const setRawSatAmount = (sats) => {
+    setSatAmount({ amount: sats, currency: WalletCurrency.BTC })
+  }
+
+  const { convertPaymentAmount } = useMySubscription()
+  const [errorMessage, setErrorMessage] = useState("")
+  const [validAmount, setValidAmount] = useState(false)
+  const usdDisabled = paymentType === "onchain" || usdWalletId === undefined
+  const isFixedAmountInvoice = fixedAmount !== undefined
+
+  useEffect(() => {
+    if (isFixedAmountInvoice) {
+      const centAmount = convertPaymentAmount(fixedAmount, WalletCurrency.USD)
+      setCentAmount(centAmount)
+    }
+  }, [convertPaymentAmount, fixedAmount, isFixedAmountInvoice])
+
+  useEffect(() => {
+    setFromWallet(
+      // Force from wallet to be BTC for onchain
+      usdDisabled
+        ? wallets.find((wallet) => wallet?.__typename === "BTCWallet")
+        : defaultWallet,
+    )
+  }, [defaultWallet, usdDisabled, wallets])
+
+  useEffect(() => {
+    if (amountCurrency === WalletCurrency.USD) {
+      setSatAmount(convertPaymentAmount(centAmount, WalletCurrency.BTC))
+    }
+    if (amountCurrency === WalletCurrency.BTC) {
+      setCentAmount(convertPaymentAmount(satAmount, WalletCurrency.USD))
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [satAmount.amount, centAmount.amount, amountCurrency, convertPaymentAmount])
+
+  useEffect(() => {
+    if (fromWallet.__typename === "BTCWallet" && satAmount) {
+      const isAmountValid = satAmount.amount <= btcWalletBalance
+      setValidAmount(isAmountValid)
+      if (isAmountValid) {
+        setErrorMessage("")
+      } else {
+        setErrorMessage(
+          translate("SendBitcoinScreen.amountExceed", {
+            balance: satAmountDisplay(btcWalletBalance),
+          }),
+        )
+      }
+    }
+    if (fromWallet.__typename === "UsdWallet" && centAmount) {
+      const isAmountValid = centAmount.amount <= usdWalletBalance
+      setValidAmount(isAmountValid)
+      if (isAmountValid) {
+        setErrorMessage("")
+      } else {
+        setErrorMessage(
+          translate("SendBitcoinScreen.amountExceed", {
+            balance: usdAmountDisplay(usdWalletBalance / 100),
+          }),
+        )
+      }
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [
+    fromWallet,
+    satAmount.amount,
+    centAmount.amount,
+    btcWalletBalance,
+    usdWalletBalance,
+  ])
+
+  if (!defaultWallet) {
+    return <></>
+  }
+
+  const toggleModal = () => {
+    setIsModalVisible(!isModalVisible)
+  }
+
+  const chooseWalletModal = (
+    <ReactNativeModal
+      style={Styles.modal}
+      animationIn="fadeInDown"
+      animationOut="fadeOutUp"
+      isVisible={isModalVisible}
+      onBackButtonPress={() => toggleModal()}
+    >
+      <View>
+        {wallets?.map((wallet) => {
+          return (
+            <TouchableWithoutFeedback
+              key={wallet.id}
+              onPress={() => {
+                setFromWallet(wallet)
+                toggleModal()
+              }}
+            >
+              <View style={Styles.fieldBackground}>
+                <View style={Styles.walletSelectorTypeContainer}>
+                  <View
+                    style={
+                      wallet.__typename === "BTCWallet"
+                        ? Styles.walletSelectorTypeLabelBitcoin
+                        : Styles.walletSelectorTypeLabelUsd
+                    }
+                  >
+                    {wallet.__typename === "BTCWallet" ? (
+                      <Text style={Styles.walletSelectorTypeLabelBtcText}>BTC</Text>
+                    ) : (
+                      <Text style={Styles.walletSelectorTypeLabelUsdText}>USD</Text>
+                    )}
+                  </View>
+                </View>
+                <View style={Styles.walletSelectorInfoContainer}>
+                  <View style={Styles.walletSelectorTypeTextContainer}>
+                    {wallet.__typename === "BTCWallet" ? (
+                      <>
+                        <Text style={Styles.walletTypeText}>Bitcoin Wallet</Text>
+                      </>
+                    ) : (
+                      <>
+                        <Text style={Styles.walletTypeText}>US Dollar Wallet</Text>
+                      </>
+                    )}
+                  </View>
+                  <View style={Styles.walletSelectorBalanceContainer}>
+                    {wallet.__typename === "BTCWallet" ? (
+                      <>
+                        <Text style={Styles.walletBalanceText}>
+                          {usdAmountDisplay(btcWalletValueInUsd)}
+                          {" - "}
+                          {satAmountDisplay(btcWalletBalance)}
+                        </Text>
+                      </>
+                    ) : (
+                      <>
+                        <Text style={Styles.walletBalanceText}>
+                          {usdAmountDisplay(usdWalletBalance / 100)}
+                        </Text>
+                      </>
+                    )}
+                  </View>
+                  <View />
+                </View>
+              </View>
+            </TouchableWithoutFeedback>
+          )
+        })}
+      </View>
+    </ReactNativeModal>
+  )
+
+  const showWalletPicker = !usdDisabled && wallets.length > 1
+
+  const goToNextScreen = async () => {
+    let invoice: string | undefined,
+      paymentAmount: PaymentAmount<WalletCurrency> | undefined
+
+    if (isFixedAmountInvoice) {
+      paymentAmount = fixedAmount
+    } else {
+      paymentAmount = amountCurrency === WalletCurrency.BTC ? satAmount : centAmount
+    }
+
+    if (paymentType === "lnurl") {
+      try {
+        const result = await fetchLnurlInvoice({
+          lnUrlOrAddress: destination,
+          tokens: satAmount.amount as Satoshis,
+        })
+        invoice = result.invoice
+      } catch (error) {
+        console.error(error)
+        setErrorMessage(translate("SendBitcoinScreen.failedToFetchLnurlInvoice"))
+        return
+      }
+    }
+
+    navigation.navigate("sendBitcoinConfirmation", {
+      lnurlInvoice: invoice,
+      fixedAmount,
+      paymentAmount,
+      recipientWalletId,
+      paymentType,
+      destination,
+      payerWalletDescriptor: {
+        id: fromWallet.id,
+        currency:
+          fromWallet.__typename === "BTCWallet" ? WalletCurrency.BTC : WalletCurrency.USD,
+      },
+      note,
+      sameNode,
+    })
+  }
+
+  return (
+    <ScrollView
+      showsVerticalScrollIndicator={false}
+      style={Styles.scrollView}
+      contentContainerStyle={Styles.contentContainer}
+    >
+      <View style={Styles.sendBitcoinAmountContainer}>
+        <View style={Styles.fieldContainer}>
+          <Text style={Styles.fieldTitleText}>{translate("common.from")}</Text>
+          <TouchableWithoutFeedback onPress={() => showWalletPicker && toggleModal()}>
+            <View style={Styles.fieldBackground}>
+              <View style={Styles.walletSelectorTypeContainer}>
+                <View
+                  style={
+                    fromWallet.__typename === "BTCWallet"
+                      ? Styles.walletSelectorTypeLabelBitcoin
+                      : Styles.walletSelectorTypeLabelUsd
+                  }
+                >
+                  {fromWallet.__typename === "BTCWallet" ? (
+                    <Text style={Styles.walletSelectorTypeLabelBtcText}>BTC</Text>
+                  ) : (
+                    <Text style={Styles.walletSelectorTypeLabelUsdText}>USD</Text>
+                  )}
+                </View>
+              </View>
+              <View style={Styles.walletSelectorInfoContainer}>
+                <View style={Styles.walletSelectorTypeTextContainer}>
+                  {fromWallet.__typename === "BTCWallet" ? (
+                    <>
+                      <Text style={Styles.walletTypeText}>Bitcoin Wallet</Text>
+                    </>
+                  ) : (
+                    <>
+                      <Text style={Styles.walletTypeText}>US Dollar Wallet</Text>
+                    </>
+                  )}
+                </View>
+                <View style={Styles.walletSelectorBalanceContainer}>
+                  {fromWallet.__typename === "BTCWallet" ? (
+                    <>
+                      <Text style={Styles.walletBalanceText}>
+                        {usdAmountDisplay(btcWalletValueInUsd)}
+                        {" - "}
+                        {satAmountDisplay(btcWalletBalance)}
+                      </Text>
+                    </>
+                  ) : (
+                    <>
+                      <Text style={Styles.walletBalanceText}>
+                        {usdAmountDisplay(usdWalletBalance / 100)}
+                      </Text>
+                    </>
+                  )}
+                </View>
+                <View />
+              </View>
+
+              {!usdDisabled && (
+                <View style={Styles.pickWalletIcon}>
+                  <Icon name={"chevron-down"} size={24} color={palette.lightBlue} />
+                </View>
+              )}
+            </View>
+          </TouchableWithoutFeedback>
+          {chooseWalletModal}
+        </View>
+        <View style={Styles.fieldContainer}>
+          <Text style={Styles.fieldTitleText}>
+            {translate("SendBitcoinScreen.amount")}
+          </Text>
+          <View style={Styles.fieldBackground}>
+            <View style={Styles.currencyInputContainer}>
+              {fromWallet.__typename === "BTCWallet" && amountCurrency === "BTC" && (
+                <>
+                  <FakeCurrencyInput
+                    value={paymentAmountToDollarsOrSats(satAmount)}
+                    onChangeValue={setRawSatAmount}
+                    prefix=""
+                    delimiter=","
+                    separator="."
+                    precision={0}
+                    suffix=" sats"
+                    minValue={0}
+                    editable={!isFixedAmountInvoice}
+                    style={Styles.walletBalanceInput}
+                  />
+                  <FakeCurrencyInput
+                    value={paymentAmountToDollarsOrSats(centAmount)}
+                    onChangeValue={setRawCentAmount}
+                    prefix="$"
+                    delimiter=","
+                    separator="."
+                    precision={2}
+                    editable={false}
+                    style={Styles.convertedAmountText}
+                  />
+                </>
+              )}
+              {fromWallet.__typename === "BTCWallet" && amountCurrency === "USD" && (
+                <>
+                  <FakeCurrencyInput
+                    value={paymentAmountToDollarsOrSats(centAmount)}
+                    onChangeValue={setRawCentAmount}
+                    prefix="$"
+                    delimiter=","
+                    separator="."
+                    precision={2}
+                    style={Styles.walletBalanceInput}
+                    minValue={0}
+                    editable={!isFixedAmountInvoice}
+                  />
+                  <FakeCurrencyInput
+                    value={paymentAmountToDollarsOrSats(satAmount)}
+                    onChangeValue={setRawSatAmount}
+                    prefix=""
+                    delimiter=","
+                    separator="."
+                    suffix=" sats"
+                    precision={0}
+                    editable={false}
+                    style={Styles.convertedAmountText}
+                  />
+                </>
+              )}
+              {fromWallet.__typename === "UsdWallet" && (
+                <FakeCurrencyInput
+                  value={paymentAmountToDollarsOrSats(centAmount)}
+                  onChangeValue={setRawCentAmount}
+                  prefix="$"
+                  delimiter=","
+                  separator="."
+                  precision={2}
+                  minValue={0}
+                  style={Styles.walletBalanceInput}
+                  editable={!isFixedAmountInvoice}
+                />
+              )}
+            </View>
+            {fromWallet.__typename === "BTCWallet" && (
+              <TouchableWithoutFeedback onPress={toggleAmountCurrency}>
+                <View style={Styles.switchCurrencyIconContainer}>
+                  <SwitchIcon />
+                </View>
+              </TouchableWithoutFeedback>
+            )}
+          </View>
+          {lnurlParams && (
+            <Text>
+              Min:{" "}
+              {fromWallet.__typename === "UsdWallet"
+                ? convertPaymentAmount(
+                    { amount: lnurlParams.min, currency: WalletCurrency.BTC },
+                    WalletCurrency.USD,
+                  ).amount / 100
+                : lnurlParams.min}{" "}
+              - Max:{" "}
+              {fromWallet.__typename === "UsdWallet"
+                ? convertPaymentAmount(
+                    { amount: lnurlParams.max, currency: WalletCurrency.BTC },
+                    WalletCurrency.USD,
+                  ).amount / 100
+                : lnurlParams.max}
+            </Text>
+          )}
+        </View>
+        <View style={Styles.fieldContainer}>
+          <Text style={Styles.fieldTitleText}>{translate("SendBitcoinScreen.note")}</Text>
+          <View style={Styles.fieldBackground}>
+            <View style={Styles.noteContainer}>
+              <View style={Styles.noteIconContainer}>
+                <NoteIcon style={Styles.noteIcon} />
+              </View>
+              <TextInput
+                style={Styles.noteInput}
+                placeholder={translate("SendBitcoinScreen.note")}
+                onChangeText={setNote}
+                value={note}
+                selectTextOnFocus
+                maxLength={lnurlParams ? lnurlParams.commentAllowed : 500}
+              />
+            </View>
+          </View>
+        </View>
+
+        {Boolean(errorMessage) && (
+          <View style={Styles.errorContainer}>
+            <Text style={Styles.errorText}>{errorMessage}</Text>
+          </View>
+        )}
+
+        <View style={Styles.buttonContainer}>
+          <Button
+            title={translate("common.next")}
+            buttonStyle={[Styles.button, Styles.activeButtonStyle]}
+            titleStyle={Styles.activeButtonTitleStyle}
+            disabledStyle={[Styles.button, Styles.disabledButtonStyle]}
+            disabledTitleStyle={Styles.disabledButtonTitleStyle}
+            disabled={!validAmount || satAmount.amount === 0}
+            onPress={goToNextScreen}
+          />
+        </View>
+      </View>
+    </ScrollView>
+  )
+}
+
+export default SendBitcoinDetailsScreen
